+++ frontend/components/text-classifier/results/classifier-annotation/ClassifierAnnotationButton.vue
